requirements.txt
torch
torchvision
hydra-core
wandb